Start renaming the app to {boQwI'}
The main reason for the separate {jIboQ} name was to make it convenient
to have a separate help page in a {jIboQ} entry in the word database.
However, since it's really not that hard to have a non-database-driven
page, we should really just have the help page exist outside of the
database for the Flutter port, and there isn't a compelling reason for
it to be named differently.