Split search page into its own StatefulWidget
The reason that search results weren't getting updated while typing
in the search query is because the widget listing the search results
was not part of the page state, so setState() had no effect. Split
the search page into its own StatefulWidget so that the results can
be properly tracked as part of the widget state. Fixes #14.