Minor help text tweaks
Put the Klingon and English full app names in separate widgets, not
just separate lines in the same widget, to prevent the lines from
being left-aligned together in a center-aligned widget. Also set
heading and subheading styles for the Klingon and English full
names of the app, respectively.

Decapitalize "keyboard" in "Klingon keyboard"