Replace menu button with back button when not on help page
It's not very obvious that swiping to the right from the left edge
of the screen is the intended way to navigate back to the previous
Navigator route on iOS. To make this essential navigation feature
more discoverable, don't add a Drawer to the Scaffold when any page
other than the help page is displayed. This will cause the Scaffold
to display the back button instead of the menu button. Users can
always access the menu from the help page; as a convenient shortcut
to return to the help page, make the title text in the AppBar push
the help page onto the Navigator when tapped.

Thanks to kechpaja for suggesting this UX improvement!