Brighten app banner slightly
Colors.red[900] corresponds to an RGB value of (200,0,0), which is
quite a bit darker than the (255,0,26) of Colors.red, which is used
for the coloration of affixes in search results. Setting the banner
to a full (255,0,26) seems a bit too bright to go with the dark
background, so as a compromise, set it to Colors.red[700], which
has an RGB value of (250,0,25), which also closer to the effective
RGB value of an affix as displayed at the body style font size,
after anti-aliasing.