Decompose 'ß' to 'ss'
This isn't a true decomposition in the Unicode sense of the word,
but it is desirable to treat 'ß' as 'ss' when searching so that
words containing 'ß' can be found whether searching with German or
Swiss German spellings (e.g. große vs. grosse).

As part of this, perform lowercasing after sanitization, to prevent
"ß" from becoming "SS" due to Klingon-casing. This also allows the
decomposition of e.g. Ö to properly become lowercased to (the
decomposed version of) ö.

Closes #22