Don't re-register URI load handler when creating a new MyHomePage
Re-registering message handler replaces the existing handler, and the
new handler will stop working if the widget that registered it is
disposed of. To avoid this, set a flag after the message handler is
registered for the root page, and don't register another handler if
the handler is already registered.

Also fix an @override flag that got misplaced in a previous change.