Allow URL links to have Klingon text names
This allows the links in the Klingon version of the help page to
display properly (pIqaD when set, or serif font when in Latin mode).
Transliterate "GitHub" as "ghItHub" ("Defense of the open hand"?)
to make it fit better, and change the ability construction (noun)
Da(verb)laH to an imperative (noun) yI(verb), since that seems more
appropriately Klingon.