Avoid null.toTile()
Following links to pre-decomposed sentences could fail to match in a
couple of cases:

* Explicit reference to all homophones with :0
* Additional tags beyond the part of speech (e.g. 'h' after homophone
  number, indicating "hidden" homophone status), which aren't part of
  the entry's key.

Attempting to resolve these components verbatim will fail to match an
entry from the word database, leading to an attempt to generate tiles
for the non-existent entries. Avoid this by handling the :0 special
case and normalizing the search name before using it as a dict key.