Explicitly decode UTF-8
Dart Strings are UTF-16, but the encoding of the JSON database is UTF-8.
Previously, multi-byte encodings seemed to be translated implicitly when
using the fromCharCodes constructor, but this doesn't seem to be the case
anymore. As a result, multi-byte UTF-8 sequences were treated as multiple
< U+0100 characters rather than the single escaped > U+007f characters
which they actually were.

Use the UTF-8 decoder instead of trying to convert the decompressed byte
stream directly into a Dart String.